Changed the default content of tests to be a call to markTestIncomplete.
I've changed the default content to something that will cause the test
to be visible in the test results rather than comments which would show
as a passing test by default.

I understand this is quite different from the initial implementation but
I think it's more robust.

I've also added another tab stop in the tests for a @covers statement
and made the methods it produces PSR1 compliant.